Reorganize shared XWiki code
* extract common helper for parsing page info responses
* put helpers for request and response handling above the Queries namespace,
  because we want to use it for commands and queries alike
