+++ modules/wikis/app/services/wikis/adapters/providers/xwiki/concerns/xwiki_page_queries.rb
@@ -32,13 +32,11 @@ module Wikis
   module Adapters
     module Providers
       module XWiki
-        module Queries
-          module Concerns
-            module XWikiPageQueries
-              def canonical_page_info(identifier:, auth_strategy:)
-                Input::PageInfo.build(identifier:).bind do |input_data|
-                  Internal::CanonicalPageInfo.new(model: provider).call(input_data:, auth_strategy:)
-                end
+        module Concerns
+          module XWikiPageQueries
+            def canonical_page_info(identifier:, auth_strategy:)
+              Input::PageInfo.build(identifier:).bind do |input_data|
+                Queries::Internal::CanonicalPageInfo.new(model: provider).call(input_data:, auth_strategy:)
               end
             end
           end
